% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstree.R
\name{mstree}
\alias{mstree}
\title{Minimum Spanning Tree}
\usage{
mstree(g)
}
\arguments{
\item{g}{The igraph object of the graph}
}
\value{
An igraph object of the MST of the graph
}
\description{
This is a function for generating the minimum spanning tree for directed or undirected graph
}
\examples{
library(igraph)
library(mygraph)
m=matrix(c(0,1,5,0,2,4,0,3,3,1,5,3,1,4,5,2,5,3,2,6,2,3,6,2,4,1,5,4,7,4,5,7,3,6,7,5)+1,ncol=3,byrow=T)
g=make_graph(t(m[,1:2]),directed=TRUE)
graph_attr(g,"weight")=m[,3]
mstree(g)

}
