% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sssp.R
\name{sssp}
\alias{sssp}
\title{Single-Source Shortest-Path}
\usage{
sssp(g, startpoint, endpoint)
}
\arguments{
\item{g}{The igraph object of the graph}

\item{startpoint}{Name of the vertex where the path starts}

\item{endpoint}{Name of the vertex where the path ends}
}
\value{
A set of vertices names of the shortest path
}
\description{
This is a function for generating the single-source shortest-path for directed or undirected graph without negative edge weights
}
\examples{
library(igraph)
library(mygraph)
m=matrix(c(0,1,5,0,2,4,0,3,3,1,5,3,1,4,5,2,5,3,2,6,2,6,3,2,4,1,5,4,7,4,5,7,3,6,7,5)+1,ncol=3,byrow=TRUE)
g1=make_graph(t(m[,1:2]),directed=TRUE)
g2=make_graph(t(m[,1:2]),directed=FALSE)
graph_attr(g1,"weight")=m[,3]
graph_attr(g2,"weight")=m[,3]
sssp(g1,1,7)
sssp(g2,1,7)
}
